Use PHP 7 versions of string handling macros.
PHP 5 has a set of string zval handling macros that require a "duplicate"
parameter that determines whether the input string will be duplicated or not.
In PHP 7, they are always duplicated. This redefines the PHP 5 macros to match
that behaviour.

This requires that the compatibility headers come in after all PHP includes,
but I think that was likely the case anyway.